+++ src/lib/Timeline.js
@@ -88,6 +88,7 @@ export default class ReactCalendarTimeline extends Component {
     keys: PropTypes.shape({
       groupIdKey: PropTypes.string,
       groupTitleKey: PropTypes.string,
+      groupLabelKey: PropTypes.string,
       groupRightTitleKey: PropTypes.string,
       itemIdKey: PropTypes.string,
       itemTitleKey: PropTypes.string,
@@ -297,7 +298,6 @@ export default class ReactCalendarTimeline extends Component {
       dragTime: null,
       dragGroupTitle: null,
       resizeTime: null,
-      topOffset: 0,
       resizingItem: null,
       resizingEdge: null
     }
@@ -415,12 +415,9 @@ export default class ReactCalendarTimeline extends Component {
   resize = (props = this.props) => {
     const {
       width: containerWidth,
-      top: containerTop
     } = this.container.getBoundingClientRect()
 
     let width = containerWidth - props.sidebarWidth - props.rightSidebarWidth
-    const { headerLabelGroupHeight, headerLabelHeight } = props
-    const headerHeight = headerLabelGroupHeight + headerLabelHeight
 
     const { dimensionItems, height, groupHeights, groupTops } = stackItems(
       props.items,
@@ -435,11 +432,9 @@ export default class ReactCalendarTimeline extends Component {
 
     // this is needed by dragItem since it uses pageY from the drag events
     // if this was in the context of the scrollElement, this would not be necessary
-    const topOffset = containerTop + window.pageYOffset + headerHeight
 
     this.setState({
       width,
-      topOffset,
       dimensionItems,
       height,
       groupHeights,
@@ -648,7 +643,7 @@ export default class ReactCalendarTimeline extends Component {
       draggingItem: item,
       dragTime: dragTime,
       newGroupOrder: newGroupOrder,
-      dragGroupTitle: newGroup ? _get(newGroup, keys.groupTitleKey) : ''
+      dragGroupTitle: newGroup ? _get(newGroup, keys.groupLabelKey) : ''
     })
   }
 
@@ -784,7 +779,6 @@ export default class ReactCalendarTimeline extends Component {
         useResizeHandle={this.props.useResizeHandle}
         canSelect={this.props.canSelect}
         moveResizeValidator={this.props.moveResizeValidator}
-        topOffset={this.state.topOffset}
         itemSelect={this.selectItem}
         itemDrag={this.dragItem}
         itemDrop={this.dropItem}
@@ -794,6 +788,7 @@ export default class ReactCalendarTimeline extends Component {
         itemResized={this.resizedItem}
         itemRenderer={this.props.itemRenderer}
         selected={this.props.selected}
+        scrollRef={this.scrollComponent}
       />
     )
   }